pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-sdlc"
version = "2.1.0"
description = "AI-powered SDLC framework with self-learning brain, automated workflows, and intelligent knowledge management"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Dao Quang Truong", email = "truongnat@gmail.com"}
]
keywords = ["ai", "sdlc", "workflow", "automation", "agent", "brain", "knowledge-management", "kit", "sdk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "PyGithub>=2.1.1",

    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "streamlit>=1.28.0",
    "PyYAML>=6.0",
    "dspy-ai>=2.1.0",
    "docker>=6.1.0",
    "googlesearch-python>=1.2.0",
    "autogen-agentchat"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
graph = [
    "neo4j>=5.14.0",
]
mcp = [
    "tavily-python>=0.3.0",
    "brave-search>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/truongnat/agentic-sdlc"
Documentation = "https://github.com/truongnat/agentic-sdlc#readme"
Repository = "https://github.com/truongnat/agentic-sdlc"
Issues = "https://github.com/truongnat/agentic-sdlc/issues"

[project.scripts]
agentic = "agentic_sdlc.cli:main"
agentic-sdlc = "agentic_sdlc.cli:main"
asdlc = "agentic_sdlc.cli:main"

[tool.setuptools.packages.find]
include = ["agentic_sdlc", "agentic_sdlc.*"]

[tool.setuptools.package-data]
agentic_sdlc = ["defaults/**/*", "requirements_tools.txt"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
